+++ src/extension.ts
@@ -5,9 +5,9 @@ import { activate as activateBuild } from './build';
 import { activate as activateDebug } from './debug';
 import { activate as activateDiagnostic } from './diagnostic';
 import { activate as activateHover } from './hover';
-import { activate as activateMinify } from './minify';
 import { activate as activateNextError } from './next-error';
 import { activate as activateRefresh } from './refresh';
+import { activate as activateTransform } from './transform';
 
 export function activate(context: ExtensionContext) {
   activateRefresh(context);
@@ -15,7 +15,7 @@ export function activate(context: ExtensionContext) {
   activateAutocomplete(context);
   activateDebug(context);
   activateBuild(context);
-  activateMinify(context);
+  activateTransform(context);
   activateNextError(context);
   activateDiagnostic(context);
 }